Possible correction re some treatment of negative numbers with absolute values in the range 2^31 to 2^62. This might have had an impact on the (ill-documented) "indirect" function or in calling foreign functions with 64-bit integer arguments. Also many fixes to serialization code as to references to shared objects, and the code in preserve & restart now changes so that the --with-test/EXPERIMENT option replaces the old code with what will eventually be the new.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3478 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arith01.cpp
@@ -304,10 +304,11 @@ int64_t sixty_four_bits(LispObject a)
             {   int len = bignum_length(a);
                 switch (len)
                 {   case CELL+4:
-                        return (int64_t)bignum_digits(a)[0]; // One word bignum
+// One word bignum. Double caset to stress sign extension.
+                        return (int64_t)(int32_t)bignum_digits(a)[0];
                     case CELL+8:
                         return bignum_digits(a)[0] |
-                               ((int64_t)bignum_digits(a)[1] << 31);
+                               ((int64_t)(int32_t)bignum_digits(a)[1] << 31);
                     default:
                         return bignum_digits(a)[0] |
                                ((int64_t)bignum_digits(a)[1] << 31) |
